Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Here's the other CL about saving sha1...

Sorry for replying late. I thought about this CL, but now have a feeling it's not helping too much with a block_list.map in TF.zip. Because in order to get that file, we have to download the entire target_files.zip anyway. And if we have the TF.zip, we can just compute the SHA-1 for the blocks in question as needed (instead of adding the build/signing time cost for always computing a bunch of SHA-1s). WDYT?

Patch-set: 7
Reviewer: Gerrit User 1056365 <1056365@85c56323-6fa9-3386-8a01-6480fb634889>