fs_config_generate add -S option
Add an -S option to fs_config_generate, that uses the supplied
-o filename as a basename instead, adding a numerical priority
prefix, and a suffix of ".system", ".vendor" or ".oem" depending
on which partition the prefix in the table has specified.  For
example if prefix has "vendor/" or "system/vendor/" then report
the output to "1-<filename>.vendor".

Add default references to the six possible
<partition>/etc/fs_config_(dirs|files) files if not specified
in the supplied or generated android_filesystem_config.h file.

Adjust make to generate with -S flag, and concatenate results into a
unified file.  This will leave intermediates for each filesystem for
future deployment.  The binary file protocol of the
fs_config_(dirs|files) files allows contatenation, and the current
orthogonal nature of the content means the order of concatenation
is unimportant, but to plan for a time when the content is not
orthogonal we concatenate in a deterministic priority order anyways.

One of those junctures may be when we have to deal with the /data
partition.  Currently the file references are informational, no tools
act upon /data references, we do not make a prepopulated /data
image, nor does adb push or sync utilize the information.  This
tool as-is automagically lands the content referencing /data into
0.fs_config_(dirs|files).system when split.  That said, we expect
/vendor/data and /oem/data could be symbolic links to /data and
that is where the differences could appear, alas the tool knows
nothing of these symbolic links.

SideEffects: none except per-partition intermediates
Test: successful build w/o -S flag (legacy behavior)
      successful build w -S flag (new behavior)
Bug: 36071012
Change-Id: I0537453bb289d41ef98ad76e69f78ae49358a1b3