fs_config_generate add -S option
Add an -S option to fs_config_generate, that uses the supplied
-o filename as a basename instead, adding a suffix of ".system",
".vendor" or ".oem" depending on which partition the prefix
in the table has specified.  For example if prefix has "vendor/" or
"system/vendor/" then report the output to "<filename>.vendor".

Add default references to the six possible
<partition>/etc/fs_config_(dirs|files) files if not specified
in the supplied or generated android_filesystem_config.h file.

Adjust make to generate with -S flag, and concatonate results
into a unified file.  This will leave intermediates for each
filesystem for future deployment.

SideEffects: none except per-partition intermediates
Test: successful build w/o -S flag (legacy behavior)
      successful build w -S flag (new behavior)
Bug: 36071012
Change-Id: I20974c2da86e9f5f306219164b7e544891cdeb2c

make

Change-Id: I0537453bb289d41ef98ad76e69f78ae49358a1b3