Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

ODM is not currently supported in the build system. Maybe we should retract the ODM portion for now?

$(if $(TARGET_OUT_ODM),$(TARGET_OUT_ODM),$(TARGET_OUT)/odm) is ok by me as an interim stage.

Also, we really need to detect if the build is using a separate image or not and tune these targets?

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1