Switch from libLLVMFuzzer to the prebuilt libFuzzer instead.
Bug: http://b/35432050

We need a match between the host prebuilt Clang and libFuzzer.a in order
to properly build/run fuzz tests. This change switches over
BUILD_FUZZ_TEST to depend on the prebuilt libFuzzer library instead. It
specifically ignores BUILD_HOST_FUZZ_TEST because we don't yet have
prebuilts for that, and there are currently no users of it.

Test: Builds successfully.
Change-Id: Ia32e36f31a8630739882ca4bed9635595adda725