Update patch set 2
Patch Set 2:

There's no choice to ignore the local notice file -- if LOCAL_NOTICE_FILE is empty, it will auto-detect the one in the current path.

We could potentially fool with the path, but that doesn't seem like the right choice.

We could also add a Soong-only flag to ignore it, but I'd still want to check to make sure that we're in the soong makefile, so this doesn't get that much simpler.

A couple more patches into this stack I also add the LLNDK stubs to this list -- those are currently easier to distinguish, since they don't have version numbers appended to themselves.

Patch-set: 2
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0