Add the ability to save off the CSV to a file. 💾
While we are generating the html file also expose a way for a user to
save off a csv file. This prevents having to run the tool twice.

* Add a new flag called csv path that takes an absolute path file to
save the info tool.
* Modify the string_for_csv and count_severity to take a writer
interface to be able to control how the content is output, as compared
to always writing to stdout.