base_rules.mk: Restrict sub-configs to the common testcase directories.
Upon further conversation and consideration, this CL restricts
sub-configs to only the common testcase directories.

The reason for this is that in the context of running CTS, having
the main config and sub-configs included would result in running
the same tests multiple times. Whereas sub-configs are meant to
group subsets of tests in suites for use in continuous infra.

Bug: 37315907
Test: `make CtsFileSystemTestCases -j` & checked the sub-config
      was only in the common testcase folder.

Change-Id: Icd099e467ea3004fa4b953a80c3c4ef17d0e9ca0