Update patch set 1
Patch Set 1:

It doesn't seem like this should be a BoardConfig option, since you're talking about needing headroom for other variants. Shouldn't this be a per-variant option? Otherwise you'd be enforcing headroom on the build variants that are tight on space.

So this should be a PRODUCT_* variable, and piped through the product*.mk files

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>