build: core: adjust the rules to convert proto files to CC file
With current build rule, to import other proto file, the
full path to the imported file should be used. To fix this
issue, the $(LOCAL_PATH) shall be specified as the first
proto_path.

With this change, the project to use proto files can follow below
rules in Android.mk:

1) If the proto files are built into a library, please make sure
   the folder "$(call local-generated-sources-dir)/proto/" is added
   into LOCAL_EXPORT_C_INCLUDE_DIRS.
2) If there is an extra ".options" file defined for the proto file,
   please use the variable LOCAL_PROTOC_OPTIMIZE_OPTIONS to refer it,
   like:
   LOCAL_PROTOC_OPTIMIZE_OPTIONS := -f $(LOCAL_PATH)/proto/sap-api.options

To include the generated header file, please just use the relative path as
same as the proto file used in Android.mk.

Test: make -j8

Change-Id: Ib57388120ad6f160ce128338a374e1014fb8546b
Signed-off-by: Xiaofeng Lei <xiaofeng.lei@ff.com>