Fix the permission issue when signing TF.zip.
sign_target_files_apks.py reconstructs recovery-from-boot.p when signing
and copying files from the input zip to the output zip. It stages the
needed files (e.g. BOOT/*, RECOVERY/*) into temporary directories for
creating the boot and recovery images.

However, symlink handling is broken after switching to soong_zip for
packaging TF.zip (soong_zip packs a symlink with 0xa1c0 instead of the
0xa1ff with regular zip). Similarly file permissions are not well
handled (it's writing with the default permission bits, which might be
different from the ones in the TF.zip). These lead to different sets of
boot/recovery images being generated (one set from the above process,
and the other from the later add_img_to_target_files.py that calls the
external unzip). As a result, it fails to update the recovery image via
/system/bin/install-recovery.sh.

This CL changes the first pass to also extract the temporary files with
the external unzip (note that we can't set the permission for symlinks
with Python's zipfile module, and os.lchmod() is not available on Linux).
This ensures that we get identical boot/recovery images from the two
passes.

Bug: 38455129
Test: Run sign_target_files_apks.py on bullhead TF.zip. Check the
      checksums in SYSTEM/bin/install-recovery.sh.
Change-Id: I1f736b90316340b77b090836c0b08cd9f8d220a0