Generate care_map.txt when AVB is enabled
When AVB is enabled, generate care_map.txt and add it to the target
files. Also copy it into the OTA package where it will later used by the
update_verifier.

Bug: 62208947
Test: Run add_img_to_target_files on the TF of a new pixel device,
and care_map.txt generates successfully. Also update_verifier succeeds
in reading all the blocks on the care_map.

Change-Id: I2881711e6f87789cb7de150dbeca18b756fed68a