make: ignore -Wmissing-noreturn warnings in clang-tidy
Now that the analyzer knows that CHECK(false) is noreturn, we're
getting clang warnings about missing noreturns in functions like:

void bar() {
  CHECK(kBarFeatureEnabled);
  /* do bar */
}

when kBarFeatureEnabled is false.

I'm unsure if this is the missing-noreturn warning being too aggressive
for its own good (e.g. if this is something we should fix in clang). For
now, though, we can just disable the warnings in clang-tidy: users will
still see the important ones when compiling their code regularly.

Bug: 32619234
Test: m. 20-something clang warnings disappear from the dashboard.
Change-Id: If579c0f53cb3ba316c105c40a1da75be17f2c7c0