Build: Change SANITIZE_LITE install settings
Under SANITIZE_LITE, the default app_process is not instrumented.
Embedding sanitized libraries into APKs will lead to crashing apps.
So move second-stage APKs to /data/asan.

The tradeoff is that for now we won't run these sanitized binaries.
Adding support for that is future work.

Also do not do a rebuild of the boot image. This would invalidate
the first-stage results. Note that this is technically dangerous,
as stack overflow guard sizes will not be adapted for ASAN runs.
However, this is a general incompatibility.

Also do not rebuild system_other. Apps are not rebuilt, so it will
create an empty image.

Bug: 36458146
Test: m && m SANITIZE_TARGET=address SANITIZE_LITE=true
Change-Id: I3898bc53cad264529f126e6bf0af9c6ca1736877
(cherry picked from commit 481660ef3419ba84e798a2659a747bf3f3e90945)