build: add TARGET_DTB_IN_SECOND
The second area can put either the 2nd stage bootloader as it was
original designed, or put a dtb.

when TARGET_DTB_IN_SECOND == true, the 2nd will be dtb;
when TARGET_NO_BOOTLOADER == false && TARGET_BOOTLOADER_IS_2ND == true,
the 2nd will be 2nd stage bootloader.

The reason why we want to reuse in the 2nd area for dtb is explained in
detail in [1].

[1] https://android-review.googlesource.com/#/c/417447/

Change-Id: I6dd295038659cf3d8a32a55d0c61a16f76929542
Signed-off-by: Bin Chen <bin.chen@linaro.org>