Update patch set 1
Patch Set 1:

Yes, I've been passing this via the command line, and it has been working, but it's been bothersome to have to do the extra typing and to have to remember to specify this property.

What would be the advantage of not having this change? Are there a significant users that run instrumentation and expect SKIP_BOOT_JARS_CHECK to be false?

I did test that explicitly setting SKIP_BOOT_JARS_CHECK=false on the command line would keep the current behavior.

Patch-set: 1
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>