Output a suite testcases's required modules/java libs.
Atest needs to be able to build and run a single test module without
building a complete suite like cts. This CL collects all of a testcases's
required modules and java libraries that also belong to a the same
suite and adds them as required dependencies of an individual test
modules.

Bug: 63295046
Test: Called make on a single test module and ensured its dependencies were
      in the host/target testcase folders.

Change-Id: Idfa9197e3ed1102d8e3779d3ffc349615f0afae2