Update patch set 1
Patch Set 1:

Soong can know whether an individual module will be called .vendor or not.  We should track the name visible to the makefiles for each module when it is generating its rules, and then use that name for the makefile dependency list.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>