Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Soong can know whether an individual module will be called .vendor or not.  We should track the name visible to the makefiles for each module when it is generating its rules, and then use that name for the makefile dependency list.

Yes, this will work, but it requires a new LOCAL_* variable just for soong-make communication. I thought the virtue of this approach was in its simplicity, but now it doesn't seem to be simpler compared to the phony module approach.

Dan, do you want me to take this path?

Patch-set: 1
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>