Update patch set 4
Patch Set 4:

Ok thanks.

I'll change the file extension to ".md". I didn't write any markdown-specific text in it yet though.

One of the advantages that I'm hoping to get from this is that the usage documentation can be versioned with the system it documents, decreasing the effort required to keep it accurate, complete, concise, and canonical.

So my plan is that the link between here and android.googlesource.com will generally come from the website and point to here (someday far in the future once the minimum relevant version has this patch) but that we might periodically export this file (with a version attached to the export) to the site too for users in case they don't want to download the source.

Patch-set: 4
Reviewer: Gerrit User 1137668 <1137668@85c56323-6fa9-3386-8a01-6480fb634889>