droiddoc: Fix trailing ':' at end of -classpath, -sourcepath
full_java_lib_deps has trailing whitespace if $(LOCAL_CLASSPATH)
is empty. Because the calculation of PRIVATE_CLASSPATH was
missing a $(strip ...), this resulted in an (unintentional)
trailing ':' in the -classpath and -sourcepath arguments to the
javadoc tool, which that tool is documented to interpret as '.'
(the working directory).

This CL adds the missing $(strip ...) so that an empty
$(LOCAL_CLASSPATH) does not inadvertently get interpreted as
the current working directory.

This issue was minor (it made no difference in practice).

Test: Treehugger
Bug: 62049770

Change-Id: Ia0e3e5657d0fa057fe998515f34bc7b8df5f6f16