Update patch set 3
Patch Set 3: Code-Review+1

> Do not have a verified +1 right in both build & build/soong to
 > trigger next step
 > 
 > BTW,currently ubsan lib seems link with prebuilt bins. Do you got a
 > plan to enable linking with fresh external/compiler-rt build libs?

external/compiler-rt is not intended for use with UBSan. UBSan is correctly using the prebuilt libraries that ship with the actual toolchain. The external/llvm (and related) repositories are actually for LLVM on-device users like RenderScript, and have nothing to do with the way that we compile regular code for the device.

Patch-set: 3
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1