Allow builds with OpenJDK 9 toolchains.
This CL allows ANDROID_JAVA_HOME to be set to a nondefault value
if EXPERIMENTAL_USE_OPENJDK9 is set.

It is an error for ANDROID_JAVA_HOME to not point to a valid toolchain
(e.g. to be empty) when EXPERIMENTAL_USE_OPENJDK9 to be set, but this
error is not explicitly checked for.

Bug: 38177295
Test: Treehugger

Change-Id: I72f641f560501e498f9c86a4380f19941fca11ad