Rename ALL_CODENAMES to ACTIVE_CODENAMES.
These aren't actually all the code names, just the ones active in our
current target. We need the list of all targets so we can validate
code names in NDK stub templates even if they're in the future, so
expose ALL_CODENAMES with an actual list of all the code names.

Test: make checkbuild
Test: check $ANDROID_PRODUCT_OUT/system/build.prop
Bug: None
Change-Id: I81fed6f761e5362ef3050be9d1e3fb17739feadd