Move default native bridge value to build.prop
Before this change it was part of default.prop, which is applied
before init.rc is executed, so that scripts cannot override the property.

In contrast, build.prop is applied at "init-late" stage,
so that one has a chance to override the property in scripts using
"init" or "init-early" triggers.

Test: fugu still works
Bug: 63133475
Change-Id: I0842abaaec1d353ebe4debda6dfe3e2c42146786