Replace ziptime with zip2zip that also sorts packages
Jars created with the jar tool are not sorted, and jar is not able
to combine -C with @ in order to take a sorted list of files.  Since
we are already running ziptime on the jar, run zip2zip instead to
both remove timestamps and sort the contents.

Also add zip2zip to compile-java to make the intermediate jars
consistent too.

Bug: 64488238
Test: m -j checkbuild
Change-Id: I97d876a2a5e26fb7c038ddc592c80034eab501ee