Fix incorrect TARGET_2ND_CPU_VARIANT for non-A/B devices
The BoardConfig.mk for arm64 non-A/B devices sets
TARGET_2ND_CPU_VARIANT to generic. This causes the generic
C implementation of memmove to be included instead of the optimized
NEON assembler implementation. This implementation is significantly
slower, in some cases enough to cause CTS failures.

The BoardConfig.mk for arm64 A/B correctly sets
TARGET_2ND_CPU_VARIANT to cortex-a15, and the comments in
build/target/board/generic_arm64/BoardConfig.mk make it quite
clear that this is the correct setting.

Bug: 65042524

Change-Id: I3aa7beddc1039b9fe3178740dfdb90858c93dff7