Fix log spam in unzip-jar-files
Commit 2bdbb6e93522b525f0322863baf9aeb5fd879d78 added
command line flags to the unzip command in unzip-jar-files
to skip the path 'module-info.class', if present.

This has led to log spam because unzip warns when an
excluded file is not present in the zip/jar file, and
that warning cannot be suppressed via command line flags.
Therefore, this CL modifies the unzip-jar-files macro
to rm -f the module-info.class file after the fact, if
it was created, which does not lead to log spam.

Test: Treehugger.
Bug: 64719206
Change-Id: Ic6be806a50557b4ba13fc18da91a8af12d14586c