Add compatibility metadata for Treble-enabled non-A/B OTA packages.
Commit 21803d3574350dbda0ba602038f8a12b1a91333d only added compatibility
metadata for A/B OTA packages, because by then time we didn't use
'ro.treble.enabled' property to determine if a target is Treble-enabled.

This CL uses 'ro.treble.enabled' to guard the packing of compatibility
metadata for both of A/B and non-A/B OTA packages. It also switches to
checking the system/vendor fingerprints to determine if there's an
update to the partition (previously it was computing the SHA-1 of the
images, which may have unintentionally changed due to issues that give
non-repetitive builds).

Bug: 64339310
Test: Generate OTA packages (full and incremental) on Treble-enabled
      target (sailfish); check that the compatibility.zip entry exists.
Test: Generate OTA packages on Treble-unenabled target (angler); check
      that the compatibility.zip entry doesn't exist.

Change-Id: I2a1fcf612439d849ba8ccea217a0faf5d5ba8e14