Skip checking files that have less blocks in block map.
When creating ext4 images with mke2fs, it may skip allocating some
blocks if they contain all zeros. As a result, there could be less
blocks listed in the block map than the actual file length. In that
case, we can't fully reconstruct a file from its block list, which in
turn fails the file consistency check. This CL skips checking for such
files to avoid failing validate_target_files.py.

Ideally this can be avoided by mounting or parsing the ext4 image
directly, which is yet to be supported in our script.

Bug: 65213616
Test: validate_target_files.py passes on targets with mke2fs generated
      images (e.g. marlin).
Change-Id: Id9cc59e345b9283844044ef94ceb5702f0ca0526