releasetools: Add --disable_streaming option.
By default we generate A/B OTA packages with streaming support since
Oreo, by carrying 'ota-streaming-property-files' in
META-INF/com/android/metadata. Note that even with this property line,
the package can still be applied in a non-streaming way, depending on
the actual server/client side config.

This CL adds '--disable_streaming' option, which allows disabling the
streaming property line generation, if we don't ever want a package to
be streamed.

Bug: 65410559
Test: Run ota_from_target_files.py w/ and w/o the option. Check
      META-INF/com/android/metadata in the generated packages.
Change-Id: Ia9063b161f395285c0786ce44f8bb62bb735fbfa