[RFC] envsetup.sh: export LC_ALL=C
The gnu sort utility is used in numerous places within the
build tree and sort ordering is dependent on LC_ALL locale
setting. Thus, enviromental factors can have an affect on
build output. Set the environment LC_ALL=C.

For instance, in some places explicit LC_ALL sets are done:
https://android.googlesource.com/platform/external/zlib/+/master/Android.mk

And in others, the enviromental default is used:
https://android.googlesource.com/platform/system/sepolicy/+/master/Android.mk#882

The environment defaults should allow for consistent builds with commands that
respect LC_ALL. Since locale of C is most sane for usages, require others
to change it if needed.

Test: **NONE** TODO DISUCSSION PATCH

Change-Id: I19dcb826f664ac293588ad23347259f791db9a8b
Signed-off-by: William Roberts <william.c.roberts@intel.com>