Update patch set 3
Patch Set 3:

That would have the same net effect but is that common?

If I am not mistaken, Jack was released as prebuilt Jars. This gives us the ability to send someone a Jar to try out and quickly revert to an old Jar.

Patch-set: 3
Reviewer: Gerrit User 1080224 <1080224@85c56323-6fa9-3386-8a01-6480fb634889>