WIP DO NOT SUBMIT: Restrict system modules to only some packages.
This CL is an exploratory experiment only and not currently intended for
submission.

After this CL, BUILD_JAVA_SYSTEM_MODULES only classes from a whitelist
of pacakges into the runtime image. For now, that whitelist corresponds
to the packages that are exported (either with or without qualification)
by OpenJDK 9's java.base.

All packages (as opposed to just those from the whitelist) are included
for core-all-system-modules, which is a runtime image used only for the
core-oj and core-libart build targets, which otherwise wouldn't compile.

Note that the classes in the whitelisted packages import classes outside
of those packages, which is forbidden on the OpenJDK 9 runtime since it
doesn't allow classes on the module path to access classes on the
classpath / in the unnamed module.

Test: Treehugger

Change-Id: I6672ee05d8ceef3e4de9b9b324fc0f1fe5e88594