Do not use RS= ORS= for awk
Removing these lines has no change on the "proper" bash build
environment, and allows using "mmm" and "mmma" on zsh.

Using RS=' ' and ORS=' ' has no real effect, and just serves to separate
entries by spaces -- which is already the default.  Using ORS=' ' also
has the unfortunate side-effect of appending a space to the final entry,
not just separating entries with spaces.

The ultimate issue is the way bash and zsh handle expansion in for loops.

Given FOO="a b c " (note trailing space) and the loop:

    for item in $FOO

bash will expand this to three entries "a", "b", and "c".

zsh will expand this into "a", "b", and "c " (note trailing space).

If the loop is modified to quote the variable:

    for item in "$FOO"

The behavior of both will be the same, with the final entry being "c ".

This means that when using "mmma directory/", the directory name is
expanded to "directory/ ", which does not exist (note that ' ' is a
valid filename for both Linux and Darwin, so this behavior is not
easily marked as "incorrect" so much as "different".

Test: bash -c 'source build/envsetup && mmma external/curl'